Refactor to install in a single transaction
This patch moves the variables to the appropriate vars folder with
the correct ansible_os_family.  It lists all of the common packages
in main.yml and distro-specific in the respective YAML file.

It also makes the installation happen with a single transaction
by providing a list to the package module.

Change-Id: Ie9b955280ef488c1e938d2925f1bdf7a7e78ea1a
Signed-off-by: Mohammed Naser <mnaser@vexxhost.com>