Update patch set 1
Patch Set 1:

The intent of this patch was to fix the fact that ansible does not work out of the box on Ubuntu systems with our Ansible scripts. This resolves that. I don't have any more time to invest in this patch so if someone wants to take it over and rewrite it with pip feel free to take it over.

Patch-set: 1