Update patch set 1
Patch Set 1:

> The intent of this patch was to fix the fact that ansible does not
 > work out of the box on Ubuntu systems with our Ansible scripts.
 > This resolves that. I don't have any more time to invest in this
 > patch so if someone wants to take it over and rewrite it with pip
 > feel free to take it over.

I'd recommend doing this as a separate task. This patch is blocking our ability to bootstrap ci-management for projects that use Ubuntu builders as their base. I think it's more critical to get this merged and resolved now so that we can unblock folks trying to use Ubuntu builders to bootstrap our CI. Can we get this merged?

Patch-set: 1