Do not install Ansible PPA on Ubuntu 20.04
Unfortunately the Ansible PPA is not yet available on Ubuntu 20.04 so
do not install it. Also add a TODO note to go back and figure out if
we really still need the PPA. This was added because we wanted to get
a more recent version of Ansible on Ubuntu 16.04 but perhaps the distro
packaged version is sufficient with newer Ubuntu versions.

Ref: https://github.com/ansible/ansible/pull/69161
Signed-off-by: Thanh Ha <zxiiro@gmail.com>
Change-Id: I7a02a6e0e95b148c640dfd41f3ea3d50dfff17d9